Bringing initial if/else check inline with runtime
I submitted a patch for Vim 7.4 a while back that modified the initial
if/else syntax check at the beginning to prevent the default CSS file
from clobbering plugin CSS syntaxes.

Since this repo seems to be the one used as the default runtime for vim,
I am syncing up the initial checks just to prevent an accidental
regression.