Fix inline/inline-block/inline-box
There was an issue where inline-block would not be detected properly due
to clashes with the keyword inline. This commit fixes that.

Also, there is no such thing as inline-table (at least that I
could find).

Also, inline-box is version of inline-block that Mozilla used in earlier
versions of Firefox.