Fix for data encoded image within cssURL
A typical data encoded image in a `url()` group will start with
something like this:

```
data:image/svg+xml;base64,...
```

The `;` character wrecks havoc on `cssAttrRegion` because it closes the
group. Using `extend` on cssURL prevents the `;` from breaking out into
`cssAttrRegion` match and killing the rest of the syntax highlighting.